zkCert: also fix padding issue outside of the snap
